+++ packages/orq-rc/RELEASES.md
@@ -599,3 +599,13 @@ Based on:
 - [python v3.2.0-rc.22] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.22] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.22 - packages/orq-rc
+
+## 2025-02-28 12:59:58
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.507.0 (2.533.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.23] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.23] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.23 - packages/orq-rc